different/requirements.txt
crsf-parser
pyserial
pygame
